+++ src/main/java/com/aqua/prod/datarest/StatusRepo.java
@@ -2,8 +2,9 @@ package com.aqua.prod.datarest;
 
 import com.aqua.prod.entity.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 
-public interface StatusRepo extends JpaRepository<Status, Long> {
+public interface StatusRepo extends JpaRepository<Status, Integer> {
     Optional<Status> getStatusByName(String name);
 }
